DatingApp/requirements.txt
django-filter==23.2
Django==4.2.2
djangorestframework==3.14.0
geopy==2.3.0
Pillow==9.5.0
psycopg2==2.9.6
python-dotenv==1.0.0